工程陪坐便椅通过

+++ studiogame/app/src/main/jni/irrlicht/CD3D8Driver.cpp
@@ -2153,7 +2153,7 @@ bool CD3D8Driver::setPixelShaderConstant(const c8* name, const f32* floats, int 
 s32 CD3D8Driver::addShaderMaterial(const c8* vertexShaderProgram,
 	const c8* pixelShaderProgram,
 	IShaderConstantSetCallBack* callback,
-	E_MATERIAL_TYPE baseMaterial, s32 userData)
+	E_MATERIAL_TYPE baseMaterial, s64 userData)
 {
 	s32 nr = -1;
 	CD3D8ShaderMaterialRenderer* r = new CD3D8ShaderMaterialRenderer(
