+++ studiogame/app/src/main/jni/irrlicht/CD3D8ParallaxMapRenderer.h
@@ -31,7 +31,7 @@ public:
 
 	//! Called by the engine when the vertex and/or pixel shader constants for an
 	//! material renderer should be set.
-	virtual void OnSetConstants(IMaterialRendererServices* services, s32 userData);
+	virtual void OnSetConstants(IMaterialRendererServices* services, s64 userData);
 
 	virtual bool OnRender(IMaterialRendererServices* service, E_VERTEX_TYPE vtxtype);
