+++ studiogame/app/src/main/jni/irrlicht/CD3D9Driver.h
@@ -338,7 +338,7 @@ namespace video
 		//! vertex shaders to render geometry.
 		s32 addShaderMaterial(const c8* vertexShaderProgram, const c8* pixelShaderProgram,
 			IShaderConstantSetCallBack* callback,
-			E_MATERIAL_TYPE baseMaterial, s32 userData);
+			E_MATERIAL_TYPE baseMaterial, s64 userData);
 
 		//! Adds a new material renderer to the VideoDriver, based on a high level shading
 		//! language.
@@ -357,7 +357,7 @@ namespace video
 			u32 verticesOut = 0,
 			IShaderConstantSetCallBack* callback = 0,
 			E_MATERIAL_TYPE baseMaterial = video::EMT_SOLID,
-			s32 userData=0);
+			s64 userData=0);
 
 		void createMaterialRenderers();
