+++ studiogame/app/src/main/jni/irrlicht/CD3D9HLSLMaterialRenderer.h
@@ -38,7 +38,7 @@ public:
 		E_PIXEL_SHADER_TYPE psCompileTarget,
 		IShaderConstantSetCallBack* callback,
 		IMaterialRenderer* baseMaterial,
-		s32 userData);
+		s64 userData);
 
 	//! Destructor
 	~CD3D9HLSLMaterialRenderer();
