+++ studiogame/app/src/main/jni/irrlicht/COGLES2ParallaxMapRenderer.h
@@ -32,7 +32,7 @@ namespace irr
 
             //! Called by the engine when the vertex and/or pixel shader constants for an
             //! material renderer should be set.
-            virtual void OnSetConstants( IMaterialRendererServices* services, s32 userData );
+            virtual void OnSetConstants( IMaterialRendererServices* services, s64 userData );
 
             virtual void OnSetMaterial( const SMaterial& material ) { }
             virtual void OnSetMaterial( const video::SMaterial& material,
