+++ studiogame/app/src/main/jni/irrlicht/COpenGLNormalMapRenderer.cpp
@@ -233,7 +233,7 @@ s32 COpenGLNormalMapRenderer::getRenderCapability() const
 
 //! Called by the engine when the vertex and/or pixel shader constants for an
 //! material renderer should be set.
-void COpenGLNormalMapRenderer::OnSetConstants(IMaterialRendererServices* services, s32 userData)
+void COpenGLNormalMapRenderer::OnSetConstants(IMaterialRendererServices* services, s64 userData)
 {
 	video::IVideoDriver* driver = services->getVideoDriver();
